Re-instated relationship between works and vehicles, succesfully
Reconnected works and vehicles so works belong to vehicles, vehicles
have many works. I just decided to forget about ownership for works

So the create controller action for works is just Work.create(work_params)
and when the vehicle_id is passed as one of those params, it correctly points
to that vehicle.

TODO: set up a "through" relationship between users and works, and delete
'records' table